@@ -20,36 +20,36 @@
 package top.theillusivec4.consecration.common.entity;
 
 import javax.annotation.Nonnull;
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.projectile.ArrowEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.world.World;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.projectile.Arrow;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 import top.theillusivec4.consecration.common.registry.ConsecrationRegistry;
 
-public class FireArrowEntity extends ArrowEntity {
+public class FireArrowEntity extends Arrow {
 
-  public FireArrowEntity(EntityType<? extends ArrowEntity> type, World world) {
+  public FireArrowEntity(EntityType<? extends Arrow> type, Level world) {
     super(type, world);
   }
 
-  public FireArrowEntity(World worldIn, double x, double y, double z) {
+  public FireArrowEntity(Level worldIn, double x, double y, double z) {
     super(worldIn, x, y, z);
   }
 
-  public FireArrowEntity(World worldIn, LivingEntity shooter) {
+  public FireArrowEntity(Level worldIn, LivingEntity shooter) {
     super(worldIn, shooter);
   }
 
   @Override
   public void tick() {
     super.tick();
-    this.setFire(100);
+    this.setSecondsOnFire(100);
   }
 
   @Override
   @Nonnull
-  protected ItemStack getArrowStack() {
+  protected ItemStack getPickupItem() {
     return new ItemStack(ConsecrationRegistry.FIRE_ARROW);
   }
 }
